1975/better seed data for ami-charts (#2156)
* fix: better seed data for ami-charts
