--- backend/core/src/seeds/ami-charts/SanJoseTCAC2019.ts
@@ -1,558 +1,0 @@
-import { AmiChartCreateDto } from "../../ami-charts/dto/ami-chart.dto"
-import { BaseEntity } from "typeorm"
-
-export const SanJoseTCAC2019: Omit<AmiChartCreateDto, keyof BaseEntity | "jurisdiction"> = {
-  name: "SanJoseTCAC2019",
-  items: [
-    {
-      percentOfAmi: 120,
-      householdSize: 1,
-      income: 110400,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 2,
-      income: 126150,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 3,
-      income: 141950,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 4,
-      income: 157700,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 5,
-      income: 170300,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 6,
-      income: 182950,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 7,
-      income: 195550,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 8,
-      income: 208150,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 1,
-      income: 101200,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 2,
-      income: 115610,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 3,
-      income: 130075,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 4,
-      income: 144540,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 5,
-      income: 156090,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 6,
-      income: 167640,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 7,
-      income: 179245,
-    },
-    {
-      percentOfAmi: 110,
-      householdSize: 8,
-      income: 190795,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 1,
-      income: 92000,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 2,
-      income: 105100,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 3,
-      income: 118250,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 4,
-      income: 131400,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 5,
-      income: 141900,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 6,
-      income: 152400,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 7,
-      income: 162950,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 8,
-      income: 173450,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 1,
-      income: 72750,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 2,
-      income: 83150,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 3,
-      income: 93550,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 4,
-      income: 103900,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 5,
-      income: 112250,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 6,
-      income: 120550,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 7,
-      income: 128850,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 8,
-      income: 137150,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 1,
-      income: 61500,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 2,
-      income: 70260,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 3,
-      income: 79020,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 4,
-      income: 87780,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 5,
-      income: 94860,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 6,
-      income: 101880,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 7,
-      income: 108900,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 8,
-      income: 115920,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 1,
-      income: 56375,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 2,
-      income: 64405,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 3,
-      income: 72435,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 4,
-      income: 80465,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 5,
-      income: 86955,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 6,
-      income: 93390,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 7,
-      income: 99825,
-    },
-    {
-      percentOfAmi: 55,
-      householdSize: 8,
-      income: 106260,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 1,
-      income: 51250,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 2,
-      income: 58550,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 3,
-      income: 65850,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 4,
-      income: 73150,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 5,
-      income: 79050,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 6,
-      income: 84900,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 7,
-      income: 90750,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 8,
-      income: 96600,
-    },
-    {
-      percentOfAmi: 45,
-      householdSize: 1,
-      income: 46125,
-    },
-    {
-      percentOfAmi: 45,
-      householdSize: 2,
-      income: 52695,
-    },
-    {
-      percentOfAmi: 45,
-      householdSize: 3,
-      income: 59265,
-    },
-    {
-      percentOfAmi: 45,
-      householdSize: 4,
-      income: 65835,
-    },
-    {
-      percentOfAmi: 45,
-      householdSize: 5,
-      income: 71145,
-    },
-    {
-      percentOfAmi: 45,
-      householdSize: 6,
-      income: 76410,
-    },
-    {
-      percentOfAmi: 45,
-      householdSize: 7,
-      income: 81675,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 1,
-      income: 41000,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 2,
-      income: 46840,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 3,
-      income: 52680,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 4,
-      income: 58520,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 5,
-      income: 63240,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 6,
-      income: 67920,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 7,
-      income: 72600,
-    },
-    {
-      percentOfAmi: 40,
-      householdSize: 8,
-      income: 77280,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 1,
-      income: 35875,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 2,
-      income: 40985,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 3,
-      income: 46095,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 4,
-      income: 51205,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 5,
-      income: 55335,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 6,
-      income: 59430,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 7,
-      income: 63525,
-    },
-    {
-      percentOfAmi: 35,
-      householdSize: 8,
-      income: 67620,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 1,
-      income: 30750,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 2,
-      income: 35130,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 3,
-      income: 39510,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 4,
-      income: 43890,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 5,
-      income: 47430,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 6,
-      income: 50940,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 7,
-      income: 54450,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 1,
-      income: 25625,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 2,
-      income: 29275,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 3,
-      income: 32925,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 4,
-      income: 36575,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 5,
-      income: 39525,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 6,
-      income: 42450,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 7,
-      income: 45375,
-    },
-    {
-      percentOfAmi: 25,
-      householdSize: 8,
-      income: 48300,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 1,
-      income: 20500,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 2,
-      income: 23420,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 3,
-      income: 26340,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 4,
-      income: 29260,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 5,
-      income: 31620,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 6,
-      income: 33960,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 7,
-      income: 36300,
-    },
-    {
-      percentOfAmi: 20,
-      householdSize: 8,
-      income: 38640,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 1,
-      income: 15375,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 2,
-      income: 17565,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 3,
-      income: 19755,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 4,
-      income: 21945,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 5,
-      income: 23715,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 6,
-      income: 25470,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 7,
-      income: 27225,
-    },
-    {
-      percentOfAmi: 15,
-      householdSize: 8,
-      income: 28980,
-    },
-  ],
-}
