--- backend/core/src/seeds/ami-charts/SanMateoCountyTCAC2020.ts
@@ -1,243 +1,0 @@
-import { AmiChartCreateDto } from "../../ami-charts/dto/ami-chart.dto"
-import { BaseEntity } from "typeorm"
-
-export const SanMateoCounty2020: Omit<AmiChartCreateDto, keyof BaseEntity | "jurisdiction"> = {
-  name: "SanMateoCounty2020",
-  items: [
-    {
-      percentOfAmi: 30,
-      householdSize: 1,
-      income: 36540,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 2,
-      income: 41760,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 3,
-      income: 46980,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 4,
-      income: 52200,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 5,
-      income: 56400,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 6,
-      income: 60570,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 7,
-      income: 64740,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 8,
-      income: 68910,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 1,
-      income: 60900,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 2,
-      income: 69600,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 3,
-      income: 78300,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 4,
-      income: 87000,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 5,
-      income: 94000,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 6,
-      income: 100950,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 7,
-      income: 107900,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 8,
-      income: 114850,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 1,
-      income: 73080,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 2,
-      income: 83520,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 3,
-      income: 93960,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 4,
-      income: 104400,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 5,
-      income: 112800,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 6,
-      income: 121140,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 7,
-      income: 129480,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 1,
-      income: 97440,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 2,
-      income: 111360,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 3,
-      income: 125280,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 4,
-      income: 139200,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 5,
-      income: 150400,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 6,
-      income: 161520,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 7,
-      income: 172640,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 8,
-      income: 183760,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 1,
-      income: 121800,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 2,
-      income: 139200,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 3,
-      income: 156600,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 4,
-      income: 174000,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 5,
-      income: 188000,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 6,
-      income: 201900,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 7,
-      income: 215800,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 8,
-      income: 229700,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 1,
-      income: 146160,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 2,
-      income: 167040,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 3,
-      income: 187920,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 4,
-      income: 208800,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 5,
-      income: 225600,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 6,
-      income: 242280,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 7,
-      income: 258960,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 8,
-      income: 275640,
-    },
-  ],
-}
