--- backend/core/src/seeds/ami-charts/SanMateoHUD2019.ts
@@ -1,233 +1,0 @@
-import { AmiChartCreateDto } from "../../ami-charts/dto/ami-chart.dto"
-import { BaseEntity } from "typeorm"
-
-export const SanMateoHUD2019: Omit<AmiChartCreateDto, keyof BaseEntity | "jurisdiction"> = {
-  name: "SanMateoHUD2019",
-  items: [
-    {
-      percentOfAmi: 120,
-      householdSize: 1,
-      income: 114900,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 2,
-      income: 131300,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 3,
-      income: 147750,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 4,
-      income: 164150,
-    },
-    {
-      percentOfAmi: 120,
-      householdSize: 5,
-      income: 177300,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 1,
-      income: 95750,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 2,
-      income: 109450,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 3,
-      income: 123100,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 4,
-      income: 136800,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 5,
-      income: 147750,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 6,
-      income: 158700,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 7,
-      income: 169650,
-    },
-    {
-      percentOfAmi: 100,
-      householdSize: 8,
-      income: 180600,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 1,
-      income: 90450,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 2,
-      income: 103350,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 3,
-      income: 116250,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 4,
-      income: 129150,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 5,
-      income: 139500,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 6,
-      income: 149850,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 7,
-      income: 160150,
-    },
-    {
-      percentOfAmi: 80,
-      householdSize: 8,
-      income: 170500,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 1,
-      income: 71170,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 2,
-      income: 81340,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 3,
-      income: 91502,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 4,
-      income: 101629,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 5,
-      income: 109833,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 6,
-      income: 117924,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 7,
-      income: 126059,
-    },
-    {
-      percentOfAmi: 60,
-      householdSize: 8,
-      income: 134219,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 1,
-      income: 56450,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 2,
-      income: 64500,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 3,
-      income: 72550,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 4,
-      income: 80600,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 5,
-      income: 87050,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 6,
-      income: 93500,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 7,
-      income: 99950,
-    },
-    {
-      percentOfAmi: 50,
-      householdSize: 8,
-      income: 106400,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 1,
-      income: 33850,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 2,
-      income: 38700,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 3,
-      income: 43550,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 4,
-      income: 48350,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 5,
-      income: 52250,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 6,
-      income: 56100,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 7,
-      income: 60000,
-    },
-    {
-      percentOfAmi: 30,
-      householdSize: 8,
-      income: 63850,
-    },
-  ],
-}
