+++ backend/core/src/seeds/ami-charts/default-ami-chart.ts
@@ -1,0 +1,315 @@
+import { InjectRepository } from "@nestjs/typeorm"
+import { Repository } from "typeorm"
+import { AmiChartCreateDto } from "../../ami-charts/dto/ami-chart.dto"
+import { AmiChart } from "../../ami-charts/entities/ami-chart.entity"
+import { Jurisdiction } from "../../jurisdictions/entities/jurisdiction.entity"
+import { CountyCode } from "../../shared/types/county-code"
+
+export function getDefaultAmiChart() {
+  return JSON.parse(JSON.stringify(defaultAmiChart))
+}
+
+export const defaultAmiChart: Omit<AmiChartCreateDto, "jurisdiction"> = {
+  name: "AlamedaCountyTCAC2021",
+  items: [
+    {
+      percentOfAmi: 80,
+      householdSize: 1,
+      income: 76720,
+    },
+    {
+      percentOfAmi: 80,
+      householdSize: 2,
+      income: 87680,
+    },
+    {
+      percentOfAmi: 80,
+      householdSize: 3,
+      income: 98640,
+    },
+    {
+      percentOfAmi: 80,
+      householdSize: 4,
+      income: 109600,
+    },
+    {
+      percentOfAmi: 80,
+      householdSize: 5,
+      income: 11840,
+    },
+    {
+      percentOfAmi: 80,
+      householdSize: 6,
+      income: 127200,
+    },
+    {
+      percentOfAmi: 80,
+      householdSize: 7,
+      income: 135920,
+    },
+    {
+      percentOfAmi: 80,
+      householdSize: 8,
+      income: 144720,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 1,
+      income: 57540,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 2,
+      income: 65760,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 3,
+      income: 73980,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 4,
+      income: 82200,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 5,
+      income: 88800,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 6,
+      income: 95400,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 7,
+      income: 101940,
+    },
+    {
+      percentOfAmi: 60,
+      householdSize: 8,
+      income: 108540,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 1,
+      income: 47950,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 2,
+      income: 54800,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 3,
+      income: 61650,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 4,
+      income: 68500,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 5,
+      income: 74000,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 6,
+      income: 79500,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 7,
+      income: 84950,
+    },
+    {
+      percentOfAmi: 50,
+      householdSize: 8,
+      income: 90450,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 1,
+      income: 43155,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 2,
+      income: 49320,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 3,
+      income: 55485,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 4,
+      income: 61650,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 5,
+      income: 66600,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 6,
+      income: 71550,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 7,
+      income: 76455,
+    },
+    {
+      percentOfAmi: 45,
+      householdSize: 8,
+      income: 81405,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 1,
+      income: 38360,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 2,
+      income: 43840,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 3,
+      income: 49320,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 4,
+      income: 54800,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 5,
+      income: 59200,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 6,
+      income: 63600,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 7,
+      income: 67960,
+    },
+    {
+      percentOfAmi: 40,
+      householdSize: 8,
+      income: 72360,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 1,
+      income: 28770,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 2,
+      income: 32880,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 3,
+      income: 36990,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 4,
+      income: 41100,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 5,
+      income: 44400,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 6,
+      income: 47700,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 7,
+      income: 50970,
+    },
+    {
+      percentOfAmi: 30,
+      householdSize: 8,
+      income: 54270,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 1,
+      income: 19180,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 2,
+      income: 21920,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 3,
+      income: 24660,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 4,
+      income: 27400,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 5,
+      income: 29600,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 6,
+      income: 31800,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 7,
+      income: 33980,
+    },
+    {
+      percentOfAmi: 20,
+      householdSize: 8,
+      income: 36180,
+    },
+  ],
+}
+
+export class AmiChartDefaultSeed {
+  constructor(
+    @InjectRepository(AmiChart)
+    protected readonly amiChartRepository: Repository<AmiChart>,
+    @InjectRepository(Jurisdiction)
+    protected readonly jurisdictionRepository: Repository<Jurisdiction>
+  ) {}
+
+  async seed() {
+    const alamedaJurisdiction = await this.jurisdictionRepository.findOneOrFail({
+      name: CountyCode.alameda,
+    })
+    return await this.amiChartRepository.save({
+      ...getDefaultAmiChart(),
+      jurisdiction: alamedaJurisdiction,
+    })
+  }
+}
