--- backend/core/src/seeds/ami-charts/index.ts
@@ -1,11 +1,0 @@
-export * from "./AlamedaCountyLIHTC2020"
-export * from "./AlamedaCountyTCAC2019"
-export * from "./AlamedaCountyTCAC2020"
-export * from "./OaklandFremontHUD2020"
-export * from "./SanJoseTCAC2019"
-export * from "./SanMateoCountyTCAC2019"
-export * from "./SanMateoCountyTCAC2020"
-export * from "./SanMateoHERASpecial2019"
-export * from "./SanMateoHOME2019"
-export * from "./SanMateoHUD2019"
-export * from "./SanMateoHUD2020"
