+++ backend/core/src/seeds/listings/listing-coliseum-seed.ts
@@ -1,7 +1,6 @@
 import { ListingSeedType, PropertySeedType, UnitSeedType } from "./listings"
 import {
   getDate,
-  getDefaultAmiChart,
   getDefaultAssets,
   getHopwaPreference,
   getLiveWorkPreference,
@@ -10,7 +9,6 @@ import {
   getTayProgram,
   PriorityTypes,
 } from "./shared"
-import { AmiChart } from "../../ami-charts/entities/ami-chart.entity"
 import { ListingStatus } from "../../listings/types/listing-status-enum"
 import { Listing } from "../../listings/entities/listing.entity"
 import { BaseEntity, DeepPartial } from "typeorm"
@@ -44,836 +42,6 @@ const coliseumProperty: PropertySeedType = {
   unitsAvailable: 46,
   yearBuilt: 2021,
 }
-const coliseumUnits: Array<UnitSeedType> = [
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "36990",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 3,
-    minOccupancy: 1,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 1,
-    number: null,
-    sqFeet: "486",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "36990",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 3,
-    minOccupancy: 1,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 1,
-    number: null,
-    sqFeet: "491",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "36990",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 3,
-    minOccupancy: 1,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 1,
-    number: null,
-    sqFeet: "491",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "61650",
-    annualIncomeMin: "38520",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 3,
-    minOccupancy: 1,
-    monthlyIncomeMin: "3210",
-    monthlyRent: "1284",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 1,
-    number: null,
-    sqFeet: "491",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "30",
-    annualIncomeMax: "44400",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 4,
-    minOccupancy: 2,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "785",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "45",
-    annualIncomeMax: "66600",
-    annualIncomeMin: "41616",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3468",
-    monthlyRent: "1387",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "45",
-    annualIncomeMax: "66600",
-    annualIncomeMin: "41616",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3468",
-    monthlyRent: "1387",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "74000",
-    annualIncomeMin: "46236",
-    bmrProgramChart: false,
-    floor: 1,
-    maxOccupancy: 5,
-    minOccupancy: 2,
-    monthlyIncomeMin: "3853",
-    monthlyRent: "1541",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 1,
-    numBedrooms: 2,
-    number: null,
-    sqFeet: "748",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "20",
-    annualIncomeMax: "31800",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "20",
-    annualIncomeMax: "31800",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 6,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1080",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "20",
-    annualIncomeMax: "31800",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "45",
-    annualIncomeMax: "71550",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "45",
-    annualIncomeMax: "71550",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "45",
-    annualIncomeMax: "71550",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "79500",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "79500",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "79500",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "79500",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "79500",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "79500",
-    annualIncomeMin: "0",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 6,
-    minOccupancy: 4,
-    monthlyIncomeMin: "0",
-    monthlyRent: null,
-    monthlyRentAsPercentOfIncome: "30",
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "84950",
-    annualIncomeMin: "53436",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 7,
-    minOccupancy: 4,
-    monthlyIncomeMin: "4453",
-    monthlyRent: "1781",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "84950",
-    annualIncomeMin: "53436",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 7,
-    minOccupancy: 4,
-    monthlyIncomeMin: "4453",
-    monthlyRent: "1781",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "84950",
-    annualIncomeMin: "53436",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 7,
-    minOccupancy: 4,
-    monthlyIncomeMin: "4453",
-    monthlyRent: "1781",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "84950",
-    annualIncomeMin: "53436",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 7,
-    minOccupancy: 4,
-    monthlyIncomeMin: "4453",
-    monthlyRent: "1781",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "84950",
-    annualIncomeMin: "53436",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 7,
-    minOccupancy: 4,
-    monthlyIncomeMin: "4453",
-    monthlyRent: "1781",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "84950",
-    annualIncomeMin: "53436",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 7,
-    minOccupancy: 4,
-    monthlyIncomeMin: "4453",
-    monthlyRent: "1781",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-  {
-    amiChart: getDefaultAmiChart() as AmiChart,
-    amiPercentage: "50",
-    annualIncomeMax: "84950",
-    annualIncomeMin: "53436",
-    bmrProgramChart: false,
-    floor: 2,
-    maxOccupancy: 7,
-    minOccupancy: 4,
-    monthlyIncomeMin: "4453",
-    monthlyRent: "1781",
-    monthlyRentAsPercentOfIncome: null,
-    numBathrooms: 2,
-    numBedrooms: 3,
-    number: null,
-    sqFeet: "1029",
-    status: UnitStatus.available,
-  },
-]
 
 const coliseumListing: ListingSeedType = {
   jurisdictionName: "Alameda",
@@ -976,14 +144,845 @@ export class ListingColiseumSeed extends ListingDefaultSeed {
     const alamedaJurisdiction = await this.jurisdictionRepository.findOneOrFail({
       name: CountyCode.alameda,
     })
-    const amiChart = await this.amiChartRepository.save({
-      ...getDefaultAmiChart(),
+    const amiChart = await this.amiChartRepository.findOneOrFail({
+      name: "AlamedaCountyTCAC2021",
       jurisdiction: alamedaJurisdiction,
     })
 
     const property = await this.propertyRepository.save({
       ...coliseumProperty,
     })
+
+    const coliseumUnits: Array<UnitSeedType> = [
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "36990",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 3,
+        minOccupancy: 1,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 1,
+        number: null,
+        sqFeet: "486",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "36990",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 3,
+        minOccupancy: 1,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 1,
+        number: null,
+        sqFeet: "491",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "36990",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 3,
+        minOccupancy: 1,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 1,
+        number: null,
+        sqFeet: "491",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "61650",
+        annualIncomeMin: "38520",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 3,
+        minOccupancy: 1,
+        monthlyIncomeMin: "3210",
+        monthlyRent: "1284",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 1,
+        number: null,
+        sqFeet: "491",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "30",
+        annualIncomeMax: "44400",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 4,
+        minOccupancy: 2,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "785",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "45",
+        annualIncomeMax: "66600",
+        annualIncomeMin: "41616",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3468",
+        monthlyRent: "1387",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "45",
+        annualIncomeMax: "66600",
+        annualIncomeMin: "41616",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3468",
+        monthlyRent: "1387",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "74000",
+        annualIncomeMin: "46236",
+        bmrProgramChart: false,
+        floor: 1,
+        maxOccupancy: 5,
+        minOccupancy: 2,
+        monthlyIncomeMin: "3853",
+        monthlyRent: "1541",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 1,
+        numBedrooms: 2,
+        number: null,
+        sqFeet: "748",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "20",
+        annualIncomeMax: "31800",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "20",
+        annualIncomeMax: "31800",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 6,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1080",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "20",
+        annualIncomeMax: "31800",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "45",
+        annualIncomeMax: "71550",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "45",
+        annualIncomeMax: "71550",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "45",
+        annualIncomeMax: "71550",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "79500",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "79500",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "79500",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "79500",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "79500",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "79500",
+        annualIncomeMin: "0",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 6,
+        minOccupancy: 4,
+        monthlyIncomeMin: "0",
+        monthlyRent: null,
+        monthlyRentAsPercentOfIncome: "30",
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "84950",
+        annualIncomeMin: "53436",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 7,
+        minOccupancy: 4,
+        monthlyIncomeMin: "4453",
+        monthlyRent: "1781",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "84950",
+        annualIncomeMin: "53436",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 7,
+        minOccupancy: 4,
+        monthlyIncomeMin: "4453",
+        monthlyRent: "1781",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "84950",
+        annualIncomeMin: "53436",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 7,
+        minOccupancy: 4,
+        monthlyIncomeMin: "4453",
+        monthlyRent: "1781",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "84950",
+        annualIncomeMin: "53436",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 7,
+        minOccupancy: 4,
+        monthlyIncomeMin: "4453",
+        monthlyRent: "1781",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "84950",
+        annualIncomeMin: "53436",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 7,
+        minOccupancy: 4,
+        monthlyIncomeMin: "4453",
+        monthlyRent: "1781",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "84950",
+        annualIncomeMin: "53436",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 7,
+        minOccupancy: 4,
+        monthlyIncomeMin: "4453",
+        monthlyRent: "1781",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+      {
+        amiChart: amiChart,
+        amiPercentage: "50",
+        annualIncomeMax: "84950",
+        annualIncomeMin: "53436",
+        bmrProgramChart: false,
+        floor: 2,
+        maxOccupancy: 7,
+        minOccupancy: 4,
+        monthlyIncomeMin: "4453",
+        monthlyRent: "1781",
+        monthlyRentAsPercentOfIncome: null,
+        numBathrooms: 2,
+        numBedrooms: 3,
+        number: null,
+        sqFeet: "1029",
+        status: UnitStatus.available,
+      },
+    ]
 
     const unitsToBeCreated: Array<Omit<UnitCreateDto, keyof BaseEntity>> = coliseumUnits.map(
       (unit) => {
