+++ backend/core/src/seeds/listings/listing-default-multiple-ami-and-percentages.ts
@@ -1,6 +1,5 @@
 import { ListingDefaultSeed } from "./listing-default-seed"
-import { getDefaultAmiChart, getDefaultProperty } from "./shared"
-import { tritonAmiChart } from "./listing-triton-seed"
+import { getDefaultProperty } from "./shared"
 import { BaseEntity } from "typeorm"
 import { UnitSeedType } from "./listings"
 import { UnitStatus } from "../../units/types/unit-status-enum"
@@ -16,12 +15,12 @@ export class ListingDefaultMultipleAMIAndPercentages extends ListingDefaultSeed 
     const alamedaJurisdiction = await this.jurisdictionRepository.findOneOrFail({
       name: CountyCode.alameda,
     })
-    const amiChartOne = await this.amiChartRepository.save({
-      ...tritonAmiChart,
+    const amiChartOne = await this.amiChartRepository.findOneOrFail({
+      name: "San Jose TCAC 2019",
       jurisdiction: alamedaJurisdiction,
     })
-    const amiChartTwo = await this.amiChartRepository.save({
-      ...getDefaultAmiChart(),
+    const amiChartTwo = await this.amiChartRepository.findOneOrFail({
+      name: "AlamedaCountyTCAC2021",
       jurisdiction: alamedaJurisdiction,
     })
